more open items moved to helper fn
